fix: removed stringify from postWithTokenAPI

+++ back-end/src/database/migrations/20230227182408-sales.js
@@ -25,7 +25,7 @@ module.exports = {
       },
       total_price: {
         allowNull: false,
-        type: Sequelize.FLOAT(9, 2),
+        type: Sequelize.DECIMAL(9, 2),
       },
       delivery_address: {
         allowNull: false,
